requirements.txt
# Telegram Bot
python-telegram-bot==21.6

# AI Clients
anthropic==0.34.2
google-generativeai==0.8.2

# Desktop Agent (local server)
fastapi==0.115.0
uvicorn==0.31.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0

# Memory (Phase 6 - optional for now)
# chromadb==0.5.5
